refactor(home): move homebody to home and remove old way to load modules
