+++ src/features/home/pages/tests/Home.native.test.tsx
@@ -5,8 +5,6 @@ import waitForExpect from 'wait-for-expect'
 import { useRoute } from '__mocks__/@react-navigation/native'
 import { UserProfileResponse } from 'api/gen'
 import { AuthContext } from 'features/auth/AuthContext'
-import { RecommendationPane, ProcessedModule } from 'features/home/contentful/moduleTypes'
-import { mockedAlgoliaResponse } from 'libs/algolia/mockedResponses/mockedAlgoliaResponse'
 import { analytics } from 'libs/analytics'
 import { env } from 'libs/environment'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
@@ -20,15 +18,6 @@ jest.mock('features/home/pages/useShowSkeleton', () => ({
 }))
 
 useRoute.mockImplementation(() => ({ params: { entryId: 'specific_entry_id' } }))
-
-let mockDisplayedModules: ProcessedModule[] = []
-jest.mock('features/home/pages/useDisplayedHomeModules', () => ({
-  useDisplayedHomeModules: jest.fn(() => ({
-    displayedModules: mockDisplayedModules,
-    homeModules: {},
-    recommendedHits: mockedAlgoliaResponse.hits.slice(0, 4),
-  })),
-}))
 
 function simulateAuthenticatedUser(partialUser?: Partial<UserProfileResponse>) {
   const user = {
@@ -176,33 +165,6 @@ describe('Home component - Analytics', () => {
     })
 
     expect(analytics.logAllModulesSeen).not.toHaveBeenCalled()
-  })
-
-  it('should trigger logEvent "RecommendationModuleSeen" when reaching the recommendation module', async () => {
-    mockDisplayedModules = [
-      new RecommendationPane({
-        display: { title: 'Tes offres recommandées', minOffers: 2, layout: 'one-item-medium' },
-      }),
-    ]
-    const home = await homeRenderer({ isLoggedIn: false })
-    const scrollView = home.getByTestId('homeBodyScrollView')
-
-    await act(async () =>
-      home
-        .getByTestId('recommendationModuleTracker')
-        .props.onLayout({ nativeEvent: { layout: { y: 1500 } } })
-    )
-    expect(home.getByTestId('recommendationModuleTracker')).toBeTruthy()
-
-    await act(async () => {
-      await scrollView.props.onScroll({ nativeEvent: nativeEventMiddle })
-    })
-    expect(analytics.logRecommendationModuleSeen).not.toHaveBeenCalled()
-
-    await act(async () => {
-      await scrollView.props.onScroll({ nativeEvent: nativeEventBottom })
-    })
-    expect(analytics.logRecommendationModuleSeen).toHaveBeenCalledWith('Tes offres recommandées', 4)
   })
 })
